Small changes requested for PR
Annotate TeslaSSOHandler and methods
Change most of logging calls to debug
Documentation changes
