+++ bundles/org.openhab.binding.tesla/src/main/java/org/openhab/binding/tesla/internal/protocol/sso/TokenResponse.java
@@ -10,18 +10,22 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.binding.tesla.internal.protocol.SSO;
+package org.openhab.binding.tesla.internal.protocol.sso;
 
 /**
- * The {@link AuthorizationCodeExchangeResponse} is a datastructure to capture
- * the response of an {@link AuthorizationCodeExchangeRequest}
+ * The {@link TokenResponse} is a datastructure to capture
+ * authentication response from Tesla Remote Service
  *
- * @author Christian Güdel
+ * @author Nicolai Grødum
  */
-public class AuthorizationCodeExchangeResponse {
+public class TokenResponse {
+
     public String access_token;
-    public String refresh_token;
-    public String expires_in;
-    public String state;
     public String token_type;
+    public Long expires_in;
+    public Long created_at;
+    public String refresh_token;
+
+    public TokenResponse() {
+    }
 }
